refact: #10 - service 인터페이스 분리 및 Integer -> int 변경
